refactor(huddle): split tts.rs tests into tts_tests.rs sibling file
tts.rs crossed the 1000-line desktop file-size budget (it was riding on
a TEMP override). Move the test module to a #[path]-included sibling —
the established pattern in this crate (commands/workflows_tests.rs,
migration_tests.rs, mesh_llm/mod_tests.rs) — and drop the override
entry. No code changes; tests are byte-identical modulo dedent.
